+++ ReleaseNotes.md
@@ -43,7 +43,7 @@ Migrated bundles from under `packages` to `bundles` and/or to new bundle-loader 
 - packages/framework/bundle/postprocessor/bundle.js -> bundles/framework/postprocessor
 - packages/framework/bundle/search/bundle.js -> bundles/framework/search
 - packages/framework/bundle/routingUI/bundle.js -> bundles/framework/routingUI
-
+- packages/framework/bundle/timeseries/bundle.js -> bundles/framework/timeseries
 - bundles/statistics/statsgrid/bundle.js -> bundles/statistics/statsgrid
 
 - packages/mapping/dimension-change/bundle.js -> bundles/mapping/dimension-change